Revert "[ESP32] Added support of door-lock for esp32 (#24169)"
This reverts commit df07c876a250c69f274e97669eaee5a2467e5cde.
