+++ examples/lock-app/esp32/main/Button.cpp
@@ -18,7 +18,7 @@
 #include "Button.h"
 #include "AppTask.h"
 
-//#include <lock/AppConfig.h>
+#include "AppConfig.h"
 
 esp_err_t Button::Init(gpio_num_t gpioNum, uint16_t debouncePeriod)
 {
