+++ examples/lock-app/esp32/main/include/AppConfig.h
@@ -17,6 +17,8 @@
 
 #pragma once
 
+#include "driver/gpio.h"
+
 // ---- Lock Example App Config ----
 
 #define APP_TASK_NAME "LOCK-APP"
