+++ examples/lock-app/esp32/main/include/AppTask.h
@@ -20,8 +20,8 @@
 #include <stdbool.h>
 #include <stdint.h>
 
-#include <lock/AppEvent.h>
-#include <lock/BoltLockManager.h>
+#include "AppEvent.h"
+#include "BoltLockManager.h"
 
 #include "freertos/FreeRTOS.h"
 #include <ble/BLEEndPoint.h>
@@ -35,7 +35,6 @@
 #define APP_ERROR_CREATE_TIMER_FAILED CHIP_APPLICATION_ERROR(0x04)
 #define APP_ERROR_START_TIMER_FAILED CHIP_APPLICATION_ERROR(0x05)
 #define APP_ERROR_STOP_TIMER_FAILED CHIP_APPLICATION_ERROR(0x06)
-#define APP_ERROR_ALLOCATION_FAILED CHIP_APPLICATION_ERROR(0x07)
 
 class AppTask
 {
