+++ src/platform/ESP32/ESP32Config.h
@@ -91,14 +91,6 @@ public:
     static const Key kConfigKey_RegulatoryLocation;
     static const Key kConfigKey_CountryCode;
     static const Key kConfigKey_UniqueId;
-    static const Key kConfigKey_LockUser;
-    static const Key kConfigKey_Credential;
-    static const Key kConfigKey_LockUserName;
-    static const Key kConfigKey_CredentialData;
-    static const Key kConfigKey_UserCredentials;
-    static const Key kConfigKey_WeekDaySchedules;
-    static const Key kConfigKey_YearDaySchedules;
-    static const Key kConfigKey_HolidaySchedules;
 
     // CHIP Counter keys
     static const Key kCounterKey_RebootCount;
